requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
# Django==5.0.1
django-filter==23.5
djangorestframework==3.14.0
Markdown==3.5.2
mysqlclient==2.2.1
pydantic==2.6.0
pydantic_core==2.16.1
pytz==2023.4
sqlparse==0.4.4
typing_extensions==4.9.0
